Horosa-Web-55c75c5b088252fbd718afeffa6d5bcb59254a0c/astropy/requirements.txt
cherrypy
jsonpickle
pyswisseph
flatlib==0.2.3.post3
